dead17: remove 13 dead functions from pkg/cli (round 2)
