+++ pkg/cli/audit_input_size_test.go
@@ -12,93 +12,6 @@ import (
 	"github.com/github/gh-aw/pkg/workflow"
 )
 
-// TestAuditReportIncludesInputSizes verifies that the audit report includes input sizes for MCP tools
-func TestAuditReportIncludesInputSizes(t *testing.T) {
-	// Create test data with input and output sizes
-	run := WorkflowRun{
-		DatabaseID:   999888,
-		WorkflowName: "MCP Tool Test",
-		Status:       "completed",
-		Conclusion:   "success",
-		CreatedAt:    time.Now(),
-		Event:        "push",
-		HeadBranch:   "main",
-		URL:          "https://github.com/test/repo/actions/runs/999888",
-		LogsPath:     "/tmp/test-logs",
-	}
-
-	metrics := LogMetrics{
-		ToolCalls: []workflow.ToolCallInfo{
-			{
-				Name:          "github_search_issues",
-				CallCount:     3,
-				MaxInputSize:  512,  // Input size in tokens
-				MaxOutputSize: 2048, // Output size in tokens
-				MaxDuration:   2 * time.Second,
-			},
-			{
-				Name:          "bash_ls",
-				CallCount:     1,
-				MaxInputSize:  64,
-				MaxOutputSize: 256,
-				MaxDuration:   500 * time.Millisecond,
-			},
-		},
-	}
-
-	processedRun := ProcessedRun{
-		Run: run,
-	}
-
-	// Generate markdown report with empty downloaded files
-	downloadedFiles := []FileInfo{}
-	report := generateAuditReport(processedRun, metrics, downloadedFiles)
-
-	// Verify the report contains the MCP Tool Usage section
-	if !strings.Contains(report, "## MCP Tool Usage") {
-		t.Error("Report should contain MCP Tool Usage section")
-	}
-
-	// Verify the table header includes "Max Input" column
-	if !strings.Contains(report, "Max Input") {
-		t.Error("Report should contain 'Max Input' column header")
-	}
-
-	// Verify input sizes are displayed
-	if !strings.Contains(report, "512") {
-		t.Error("Report should contain input size 512")
-	}
-	if !strings.Contains(report, "64") {
-		t.Error("Report should contain input size 64")
-	}
-
-	// Verify output sizes are still displayed
-	if !strings.Contains(report, "2.05k") && !strings.Contains(report, "2.0k") && !strings.Contains(report, "2048") {
-		t.Error("Report should contain output size 2.05k, 2.0k or 2048")
-	}
-
-	// Verify durations are displayed
-	if !strings.Contains(report, "2.0s") && !strings.Contains(report, "2s") {
-		t.Error("Report should contain duration 2.0s or 2s")
-	}
-
-	// Verify the table has the correct number of columns (5: Tool, Calls, Max Input, Max Output, Max Duration)
-	headerLine := ""
-	for line := range strings.SplitSeq(report, "\n") {
-		if strings.Contains(line, "Max Input") {
-			headerLine = line
-			break
-		}
-	}
-	if headerLine != "" {
-		// Count pipes to verify we have the right number of columns
-		pipeCount := strings.Count(headerLine, "|")
-		if pipeCount != 6 { // 6 pipes for 5 columns (includes leading and trailing)
-			t.Errorf("Expected 6 pipes in header line, got %d. Line: %s", pipeCount, headerLine)
-		}
-	}
-}
-
 // TestAuditDataJSONIncludesInputSizes verifies that JSON output includes input sizes
 func TestAuditDataJSONIncludesInputSizes(t *testing.T) {
 	run := WorkflowRun{
