+++ pkg/cli/audit_test.go
@@ -17,95 +17,6 @@ import (
 	"github.com/github/gh-aw/pkg/workflow"
 )
 
-func TestExtractRunID(t *testing.T) {
-	tests := []struct {
-		name      string
-		input     string
-		expected  int64
-		shouldErr bool
-	}{
-		{
-			name:      "Numeric run ID",
-			input:     "1234567890",
-			expected:  1234567890,
-			shouldErr: false,
-		},
-		{
-			name:      "Run URL",
-			input:     "https://github.com/owner/repo/actions/runs/12345678",
-			expected:  12345678,
-			shouldErr: false,
-		},
-		{
-			name:      "Job URL",
-			input:     "https://github.com/owner/repo/actions/runs/12345678/job/98765432",
-			expected:  12345678,
-			shouldErr: false,
-		},
-		{
-			name:      "Job URL with attempts",
-			input:     "https://github.com/owner/repo/actions/runs/12345678/attempts/2",
-			expected:  12345678,
-			shouldErr: false,
-		},
-		{
-			name:      "Run URL with trailing slash",
-			input:     "https://github.com/owner/repo/actions/runs/12345678/",
-			expected:  12345678,
-			shouldErr: false,
-		},
-		{
-			name:      "Workflow run URL without /actions/",
-			input:     "https://github.com/owner/repo/runs/12345678",
-			expected:  12345678,
-			shouldErr: false,
-		},
-		{
-			name:      "GitHub Enterprise URL",
-			input:     "https://github.example.com/owner/repo/actions/runs/12345678",
-			expected:  12345678,
-			shouldErr: false,
-		},
-		{
-			name:      "Invalid format",
-			input:     "not-a-number",
-			expected:  0,
-			shouldErr: true,
-		},
-		{
-			name:      "Invalid URL without run ID",
-			input:     "https://github.com/owner/repo/actions",
-			expected:  0,
-			shouldErr: true,
-		},
-		{
-			name:      "Empty string",
-			input:     "",
-			expected:  0,
-			shouldErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result, err := extractRunID(tt.input)
-
-			if tt.shouldErr {
-				if err == nil {
-					t.Errorf("Expected error but got none")
-				}
-			} else {
-				if err != nil {
-					t.Errorf("Unexpected error: %v", err)
-				}
-				if result != tt.expected {
-					t.Errorf("Expected run ID %d, got %d", tt.expected, result)
-				}
-			}
-		})
-	}
-}
-
 func TestParseRunURL(t *testing.T) {
 	tests := []struct {
 		name             string
@@ -261,318 +172,6 @@ func TestIsPermissionError(t *testing.T) {
 				t.Errorf("Expected %v, got %v", tt.expected, result)
 			}
 		})
-	}
-}
-
-func TestGenerateAuditReport(t *testing.T) {
-	// Create test data
-	run := WorkflowRun{
-		DatabaseID:    123456,
-		WorkflowName:  "Test Workflow",
-		Status:        "completed",
-		Conclusion:    "success",
-		CreatedAt:     time.Date(2024, 1, 1, 10, 0, 0, 0, time.UTC),
-		StartedAt:     time.Date(2024, 1, 1, 10, 0, 30, 0, time.UTC),
-		UpdatedAt:     time.Date(2024, 1, 1, 10, 5, 0, 0, time.UTC),
-		Duration:      4*time.Minute + 30*time.Second,
-		Event:         "push",
-		HeadBranch:    "main",
-		URL:           "https://github.com/org/repo/actions/runs/123456",
-		TokenUsage:    1500,
-		EstimatedCost: 0.025,
-		Turns:         5,
-		ErrorCount:    0,
-		WarningCount:  1,
-		LogsPath:      "/tmp/gh-aw/test-logs",
-	}
-
-	metrics := LogMetrics{
-		TokenUsage:    1500,
-		EstimatedCost: 0.025,
-		Turns:         5,
-		ToolCalls: []workflow.ToolCallInfo{
-			{
-				Name:          "github_issue_read",
-				CallCount:     3,
-				MaxInputSize:  256,
-				MaxOutputSize: 1024,
-				MaxDuration:   2 * time.Second,
-			},
-			{
-				Name:          "bash_echo",
-				CallCount:     2,
-				MaxInputSize:  128,
-				MaxOutputSize: 512,
-				MaxDuration:   1 * time.Second,
-			},
-		},
-	}
-
-	missingTools := []MissingToolReport{
-		{
-			Tool:         "missing_tool",
-			Reason:       "Tool not available",
-			Alternatives: "use alternative_tool instead",
-			Timestamp:    "2024-01-01T10:00:00Z",
-		},
-	}
-
-	mcpFailures := []MCPFailureReport{
-		{
-			ServerName: "test-server",
-			Status:     "failed",
-		},
-	}
-
-	processedRun := ProcessedRun{
-		Run:          run,
-		MissingTools: missingTools,
-		MCPFailures:  mcpFailures,
-	}
-
-	// Create test downloaded files
-	downloadedFiles := []FileInfo{
-		{
-			Path:        "aw_info.json",
-			Size:        256,
-			Description: "Engine configuration and workflow metadata",
-		},
-		{
-			Path:        "safe_output.jsonl",
-			Size:        512,
-			Description: "Safe outputs from workflow execution",
-		},
-	}
-
-	// Generate report
-	report := generateAuditReport(processedRun, metrics, downloadedFiles)
-
-	// Verify report contains expected sections
-	expectedSections := []string{
-		"# Workflow Run Audit Report",
-		"## Overview",
-		"## Metrics",
-		"## MCP Tool Usage",
-		"## MCP Server Failures",
-		"## Missing Tools",
-		"## Downloaded Files",
-	}
-
-	for _, section := range expectedSections {
-		if !strings.Contains(report, section) {
-			t.Errorf("Report missing expected section: %s", section)
-		}
-	}
-
-	// Verify report contains specific data
-	expectedContent := []string{
-		"123456",        // Run ID
-		"Test Workflow", // Workflow name
-		"success",       // Conclusion
-		"main",          // Branch
-		"0.025",         // Estimated cost
-		"5",             // Turns
-		"missing_tool",  // Missing tool
-		"test-server",   // MCP failure
-	}
-
-	for _, content := range expectedContent {
-		if !strings.Contains(report, content) {
-			t.Errorf("Report missing expected content: %s", content)
-		}
-	}
-
-	// Token usage should be present (formatted as 1.5k or similar)
-	if !strings.Contains(report, "1.5k") && !strings.Contains(report, "1500") && !strings.Contains(report, "Token Usage") {
-		t.Errorf("Report missing token usage (should be 1.5k or 1500)\nReport:\n%s", report)
-	}
-}
-
-func TestGenerateAuditReportMinimal(t *testing.T) {
-	// Test with minimal data (no errors, no tools, etc.)
-	run := WorkflowRun{
-		DatabaseID:   789,
-		WorkflowName: "Minimal Workflow",
-		Status:       "in_progress",
-		CreatedAt:    time.Now(),
-		Event:        "workflow_dispatch",
-		HeadBranch:   "feature",
-		URL:          "https://github.com/org/repo/actions/runs/789",
-		LogsPath:     "/tmp/gh-aw/minimal-logs",
-	}
-
-	metrics := LogMetrics{}
-
-	processedRun := ProcessedRun{
-		Run: run,
-	}
-
-	// Generate report with empty downloaded files
-	downloadedFiles := []FileInfo{}
-	report := generateAuditReport(processedRun, metrics, downloadedFiles)
-
-	// Verify report contains basic sections even with minimal data
-	expectedSections := []string{
-		"# Workflow Run Audit Report",
-		"## Overview",
-		"## Metrics",
-		"## Downloaded Files",
-	}
-
-	for _, section := range expectedSections {
-		if !strings.Contains(report, section) {
-			t.Errorf("Minimal report missing expected section: %s", section)
-		}
-	}
-
-	// Verify it doesn't contain sections that should be omitted when empty
-	unexpectedSections := []string{
-		"## MCP Server Failures",
-		"## Missing Tools",
-		"## Issue Summary",
-	}
-
-	for _, section := range unexpectedSections {
-		if strings.Contains(report, section) {
-			t.Errorf("Minimal report should not contain section: %s", section)
-		}
-	}
-}
-
-func TestGenerateAuditReportWithErrors(t *testing.T) {
-	// Test with errors to verify issue summary
-	run := WorkflowRun{
-		DatabaseID:   999,
-		WorkflowName: "Error Workflow",
-		Status:       "completed",
-		Conclusion:   "failure",
-		CreatedAt:    time.Now(),
-		Event:        "push",
-		HeadBranch:   "main",
-		URL:          "https://github.com/org/repo/actions/runs/999",
-		ErrorCount:   3,
-		WarningCount: 2,
-		LogsPath:     "/tmp/gh-aw/error-logs",
-	}
-
-	metrics := LogMetrics{}
-
-	processedRun := ProcessedRun{
-		Run: run,
-	}
-
-	// Generate report with empty downloaded files
-	downloadedFiles := []FileInfo{}
-	report := generateAuditReport(processedRun, metrics, downloadedFiles)
-
-	// Verify issue summary is present
-	if !strings.Contains(report, "## Issue Summary") {
-		t.Error("Report should contain Issue Summary when errors are present")
-	}
-
-	// Verify error counts are mentioned
-	if !strings.Contains(report, "3 error(s)") {
-		t.Error("Report should mention error count")
-	}
-	if !strings.Contains(report, "2 warning(s)") {
-		t.Error("Report should mention warning count")
-	}
-
-	// Note: Individual error/warning extraction was removed from buildAuditData
-	// The errors/warnings section generation and individual error display
-	// is no longer performed
-	// if !strings.Contains(report, "### Errors and Warnings") {
-	// 	t.Error("Report should contain 'Errors and Warnings' section")
-	// }
-	// if !strings.Contains(report, "Failed to initialize tool") {
-	// 	t.Error("Report should contain first error message")
-	// }
-	// if !strings.Contains(report, "Connection timeout") {
-	// 	t.Error("Report should contain second error message")
-	// }
-	// if !strings.Contains(report, "Deprecated API usage") {
-	// 	t.Error("Report should contain warning message")
-	// }
-	// if !strings.Contains(report, "agent.log:10:") {
-	// 	t.Error("Report should contain file:line format for first error")
-	// }
-}
-
-func TestGenerateAuditReportArtifacts(t *testing.T) {
-	// Create temporary directory with test artifacts
-	tmpDir := testutil.TempDir(t, "test-*")
-
-	// Create test artifact files
-	artifacts := []string{
-		"aw_info.json",
-		"safe_output.jsonl",
-		"aw.patch",
-		"agent_output.json",
-	}
-
-	for _, artifact := range artifacts {
-		if err := os.WriteFile(filepath.Join(tmpDir, artifact), []byte("test"), 0644); err != nil {
-			t.Fatalf("Failed to create test artifact %s: %v", artifact, err)
-		}
-	}
-
-	run := WorkflowRun{
-		DatabaseID:   555,
-		WorkflowName: "Artifact Test",
-		Status:       "completed",
-		Conclusion:   "success",
-		CreatedAt:    time.Now(),
-		Event:        "push",
-		HeadBranch:   "main",
-		URL:          "https://github.com/org/repo/actions/runs/555",
-		LogsPath:     tmpDir,
-	}
-
-	metrics := LogMetrics{}
-
-	processedRun := ProcessedRun{
-		Run: run,
-	}
-
-	// Create test downloaded files matching the artifacts
-	downloadedFiles := []FileInfo{
-		{
-			Path:        "aw_info.json",
-			Size:        4,
-			Description: "Engine configuration and workflow metadata",
-		},
-		{
-			Path:        "safe_output.jsonl",
-			Size:        4,
-			Description: "Safe outputs from workflow execution",
-		},
-		{
-			Path:        "aw.patch",
-			Size:        4,
-			Description: "Git patch of changes made during execution",
-		},
-		{
-			Path:        "agent_output.json",
-			Size:        4,
-			Description: "Validated safe outputs",
-		},
-	}
-
-	// Generate report
-	report := generateAuditReport(processedRun, metrics, downloadedFiles)
-
-	// Verify all artifacts are listed
-	expectedArtifacts := []string{
-		"aw_info.json",
-		"safe_output.jsonl",
-		"aw.patch",
-		"agent_output.json",
-	}
-
-	for _, artifact := range expectedArtifacts {
-		if !strings.Contains(report, artifact) {
-			t.Errorf("Report should list artifact: %s", artifact)
-		}
 	}
 }
 
@@ -1006,90 +605,6 @@ func TestBuildAuditDataWithFirewall(t *testing.T) {
 	}
 	if len(auditData.FirewallAnalysis.BlockedDomains) != 1 {
 		t.Errorf("Expected 1 blocked domain, got %d", len(auditData.FirewallAnalysis.BlockedDomains))
-	}
-}
-
-func TestGenerateAuditReportWithFirewall(t *testing.T) {
-	// Create test data with firewall analysis
-	run := WorkflowRun{
-		DatabaseID:    123456,
-		WorkflowName:  "Test Workflow",
-		Status:        "completed",
-		Conclusion:    "success",
-		CreatedAt:     time.Date(2024, 1, 1, 10, 0, 0, 0, time.UTC),
-		Event:         "push",
-		HeadBranch:    "main",
-		URL:           "https://github.com/org/repo/actions/runs/123456",
-		TokenUsage:    1500,
-		EstimatedCost: 0.025,
-		Turns:         5,
-		ErrorCount:    0,
-		WarningCount:  0,
-		LogsPath:      "/tmp/gh-aw/test-logs",
-	}
-
-	metrics := LogMetrics{
-		TokenUsage:    1500,
-		EstimatedCost: 0.025,
-		Turns:         5,
-	}
-
-	firewallAnalysis := &FirewallAnalysis{
-		DomainBuckets: DomainBuckets{
-			AllowedDomains: []string{"api.github.com:443", "npmjs.org:443"},
-			BlockedDomains: []string{"blocked.example.com:443"},
-		},
-		TotalRequests:   10,
-		AllowedRequests: 7,
-		BlockedRequests: 3,
-		RequestsByDomain: map[string]DomainRequestStats{
-			"api.github.com:443":      {Allowed: 5, Blocked: 0},
-			"npmjs.org:443":           {Allowed: 2, Blocked: 0},
-			"blocked.example.com:443": {Allowed: 0, Blocked: 3},
-		},
-	}
-
-	processedRun := ProcessedRun{
-		Run:              run,
-		FirewallAnalysis: firewallAnalysis,
-		MissingTools:     []MissingToolReport{},
-		MCPFailures:      []MCPFailureReport{},
-	}
-
-	// Generate report with empty downloaded files
-	downloadedFiles := []FileInfo{}
-	report := generateAuditReport(processedRun, metrics, downloadedFiles)
-
-	// Verify firewall section is present
-	if !strings.Contains(report, "## Firewall Analysis") {
-		t.Error("Report should contain Firewall Analysis section")
-	}
-
-	// Verify firewall statistics are mentioned
-	if !strings.Contains(report, "Total Requests") {
-		t.Error("Report should mention total requests")
-	}
-	if !strings.Contains(report, "Allowed Requests") {
-		t.Error("Report should mention allowed requests")
-	}
-	if !strings.Contains(report, "Blocked Requests") {
-		t.Error("Report should mention blocked requests")
-	}
-
-	// Verify domain lists are present
-	if !strings.Contains(report, "Allowed Domains") {
-		t.Error("Report should contain Allowed Domains section")
-	}
-	if !strings.Contains(report, "Blocked Domains") {
-		t.Error("Report should contain Blocked Domains section")
-	}
-
-	// Verify specific domains are listed
-	if !strings.Contains(report, "api.github.com:443") {
-		t.Error("Report should list allowed domain api.github.com:443")
-	}
-	if !strings.Contains(report, "blocked.example.com:443") {
-		t.Error("Report should list blocked domain blocked.example.com:443")
 	}
 }
