+++ pkg/cli/commands_file_watching_test.go
@@ -250,78 +250,6 @@ func TestCompileAllWorkflowFiles(t *testing.T) {
 	})
 }
 
-// TestCompileModifiedFiles tests the compileModifiedFiles function
-// This covers pkg/cli/commands.go:837
-func TestCompileModifiedFiles(t *testing.T) {
-	t.Run("compile modified files basic functionality", func(t *testing.T) {
-		tempDir := testutil.TempDir(t, "test-*")
-		workflowsDir := filepath.Join(tempDir, ".github/workflows")
-		os.MkdirAll(workflowsDir, 0755)
-
-		// Create test files with different modification times
-		file1 := filepath.Join(workflowsDir, "recent.md")
-		file2 := filepath.Join(workflowsDir, "old.md")
-
-		content := "---\non: push\nengine: claude\n---\n# Test\n\nTest content"
-
-		os.WriteFile(file1, []byte(content), 0644)
-		os.WriteFile(file2, []byte(content), 0644)
-
-		// Make file2 older
-		oldTime := time.Now().Add(-2 * time.Hour)
-		os.Chtimes(file2, oldTime, oldTime)
-
-		compiler := workflow.NewCompiler()
-
-		// Test with recent files - compileModifiedFiles takes a slice of files
-		modifiedFiles := []string{file1} // Only include the recent file
-		compileModifiedFiles(compiler, modifiedFiles, true)
-
-		// Check that the recent file was compiled
-		recentLock := filepath.Join(workflowsDir, "recent.lock.yml")
-
-		if _, err := os.Stat(recentLock); os.IsNotExist(err) {
-			t.Error("Recent file should have been compiled")
-		}
-	})
-
-	t.Run("compile modified files with no files", func(t *testing.T) {
-		compiler := workflow.NewCompiler()
-
-		// Test with empty file list (should not error)
-		emptyFiles := []string{}
-		compileModifiedFiles(compiler, emptyFiles, true)
-		// Should complete without error
-	})
-
-	t.Run("compile modified files with invalid files", func(t *testing.T) {
-		compiler := workflow.NewCompiler()
-
-		// Test with invalid file paths
-		invalidFiles := []string{"nonexistent/path/file.md"}
-		compileModifiedFiles(compiler, invalidFiles, false)
-		// Should handle gracefully without panicking
-	})
-
-	t.Run("compile modified files verbose mode", func(t *testing.T) {
-		tempDir := testutil.TempDir(t, "test-*")
-		workflowsDir := filepath.Join(tempDir, ".github/workflows")
-		os.MkdirAll(workflowsDir, 0755)
-
-		// Create a recent file
-		recentFile := filepath.Join(workflowsDir, "recent.md")
-		content := "---\non: push\nengine: claude\n---\n# Recent Test\n\nRecent content"
-		os.WriteFile(recentFile, []byte(content), 0644)
-
-		compiler := workflow.NewCompiler()
-
-		// Test verbose mode
-		modifiedFiles := []string{recentFile}
-		compileModifiedFiles(compiler, modifiedFiles, true)
-		// Should complete without error
-	})
-}
-
 // TestHandleFileDeleted tests the handleFileDeleted function
 // This covers pkg/cli/commands.go:888
 func TestHandleFileDeleted(t *testing.T) {
