+++ pkg/cli/copilot_setup.go
@@ -315,60 +315,3 @@ func upgradeSetupCliVersion(workflow *Workflow, actionMode workflow.ActionMode, 
 
 	return upgraded, nil
 }
-
-// injectExtensionInstallStep injects the gh-aw extension install and verification steps into an existing workflow
-func injectExtensionInstallStep(workflow *Workflow, actionMode workflow.ActionMode, version string) error {
-	var installStep, checkoutStep CopilotWorkflowStep
-
-	// Determine the action reference - use version tag in release mode, @main in dev mode
-	actionRef := getActionRef(actionMode, version)
-
-	if actionMode.IsRelease() {
-		// In release mode, use the actions/setup-cli action
-		checkoutStep = CopilotWorkflowStep{
-			Name: "Checkout repository",
-			Uses: "actions/checkout@v6",
-		}
-		installStep = CopilotWorkflowStep{
-			Name: "Install gh-aw extension",
-			Uses: "github/gh-aw/actions/setup-cli" + actionRef,
-			With: map[string]any{
-				"version": version,
-			},
-		}
-	} else {
-		// In dev/script mode, use curl to download install script
-		installStep = CopilotWorkflowStep{
-			Name: "Install gh-aw extension",
-			Run:  "curl -fsSL https://raw.githubusercontent.com/github/gh-aw/refs/heads/main/install-gh-aw.sh | bash",
-		}
-	}
-
-	// Find the copilot-setup-steps job
-	job, exists := workflow.Jobs["copilot-setup-steps"]
-	if !exists {
-		return errors.New("copilot-setup-steps job not found in workflow")
-	}
-
-	// Insert the extension install step at the beginning
-	insertPosition := 0
-
-	// Prepare steps to insert based on mode
-	var stepsToInsert []CopilotWorkflowStep
-	if actionMode.IsRelease() {
-		stepsToInsert = []CopilotWorkflowStep{checkoutStep, installStep}
-	} else {
-		stepsToInsert = []CopilotWorkflowStep{installStep}
-	}
-
-	// Insert steps at the determined position
-	newSteps := make([]CopilotWorkflowStep, 0, len(job.Steps)+len(stepsToInsert))
-	newSteps = append(newSteps, job.Steps[:insertPosition]...)
-	newSteps = append(newSteps, stepsToInsert...)
-	newSteps = append(newSteps, job.Steps[insertPosition:]...)
-
-	job.Steps = newSteps
-	workflow.Jobs["copilot-setup-steps"] = job
-
-	return nil
-}
