+++ pkg/cli/copilot_setup_test.go
@@ -177,98 +177,6 @@ func TestEnsureCopilotSetupSteps(t *testing.T) {
 	}
 }
 
-func TestInjectExtensionInstallStep(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name          string
-		workflow      *Workflow
-		wantErr       bool
-		expectedSteps int
-		validateFunc  func(*testing.T, *Workflow)
-	}{
-		{
-			name: "injects at beginning of existing steps",
-			workflow: &Workflow{
-				Jobs: map[string]WorkflowJob{
-					"copilot-setup-steps": {
-						Steps: []CopilotWorkflowStep{
-							{Name: "Some step"},
-							{Name: "Build"},
-						},
-					},
-				},
-			},
-			wantErr:       false,
-			expectedSteps: 3, // 1 injected step + 2 existing steps
-			validateFunc: func(t *testing.T, w *Workflow) {
-				steps := w.Jobs["copilot-setup-steps"].Steps
-				// Extension install should be at index 0 (beginning)
-				if steps[0].Name != "Install gh-aw extension" {
-					t.Errorf("Expected step 0 to be 'Install gh-aw extension', got %q", steps[0].Name)
-				}
-			},
-		},
-		{
-			name: "injects when no existing steps",
-			workflow: &Workflow{
-				Jobs: map[string]WorkflowJob{
-					"copilot-setup-steps": {
-						Steps: []CopilotWorkflowStep{},
-					},
-				},
-			},
-			wantErr:       false,
-			expectedSteps: 1, // 1 injected step (install only)
-			validateFunc: func(t *testing.T, w *Workflow) {
-				steps := w.Jobs["copilot-setup-steps"].Steps
-				// Should have 1 step
-				if len(steps) != 1 {
-					t.Errorf("Expected 1 step, got %d", len(steps))
-				}
-				if steps[0].Name != "Install gh-aw extension" {
-					t.Errorf("Expected step 0 to be 'Install gh-aw extension', got %q", steps[0].Name)
-				}
-			},
-		},
-		{
-			name: "returns error when job not found",
-			workflow: &Workflow{
-				Jobs: map[string]WorkflowJob{
-					"other-job": {
-						Steps: []CopilotWorkflowStep{},
-					},
-				},
-			},
-			wantErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := injectExtensionInstallStep(tt.workflow, workflow.ActionModeDev, "dev")
-
-			if (err != nil) != tt.wantErr {
-				t.Errorf("injectExtensionInstallStep() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if tt.wantErr {
-				return
-			}
-
-			job := tt.workflow.Jobs["copilot-setup-steps"]
-			if len(job.Steps) != tt.expectedSteps {
-				t.Errorf("Expected %d steps, got %d", tt.expectedSteps, len(job.Steps))
-			}
-
-			if tt.validateFunc != nil {
-				tt.validateFunc(t, tt.workflow)
-			}
-		})
-	}
-}
-
 func TestWorkflowStructMarshaling(t *testing.T) {
 	t.Parallel()
 
@@ -936,98 +844,6 @@ jobs:
 
 	if string(content) != existingContent {
 		t.Errorf("Expected file to remain unchanged")
-	}
-}
-
-// TestInjectExtensionInstallStep_ReleaseMode tests injecting in release mode with version
-func TestInjectExtensionInstallStep_ReleaseMode(t *testing.T) {
-	wf := &Workflow{
-		Jobs: map[string]WorkflowJob{
-			"copilot-setup-steps": {
-				Steps: []CopilotWorkflowStep{
-					{Name: "Existing step", Run: "echo test"},
-				},
-			},
-		},
-	}
-
-	testVersion := "v4.5.6"
-	err := injectExtensionInstallStep(wf, workflow.ActionModeRelease, testVersion)
-	if err != nil {
-		t.Fatalf("injectExtensionInstallStep() failed: %v", err)
-	}
-
-	job := wf.Jobs["copilot-setup-steps"]
-
-	// Should have 3 steps: checkout, install, existing (no verify)
-	if len(job.Steps) != 3 {
-		t.Fatalf("Expected 3 steps (checkout, install, existing), got %d", len(job.Steps))
-	}
-
-	// Verify checkout step
-	if job.Steps[0].Name != "Checkout repository" {
-		t.Errorf("First step should be checkout, got: %s", job.Steps[0].Name)
-	}
-	if job.Steps[0].Uses != "actions/checkout@v6" {
-		t.Errorf("Checkout should use actions/checkout@v6, got: %s", job.Steps[0].Uses)
-	}
-
-	// Verify install step
-	if job.Steps[1].Name != "Install gh-aw extension" {
-		t.Errorf("Second step should be install, got: %s", job.Steps[1].Name)
-	}
-	expectedUses := "github/gh-aw/actions/setup-cli@v4.5.6"
-	if job.Steps[1].Uses != expectedUses {
-		t.Errorf("Install should use %s, got: %s", expectedUses, job.Steps[1].Uses)
-	}
-	if version, ok := job.Steps[1].With["version"]; !ok || version != testVersion {
-		t.Errorf("Install step should have version: %s in with, got: %v", testVersion, job.Steps[1].With)
-	}
-
-	// Verify original step is preserved
-	if job.Steps[2].Name != "Existing step" {
-		t.Errorf("Third step should be existing step, got: %s", job.Steps[2].Name)
-	}
-}
-
-// TestInjectExtensionInstallStep_DevMode tests injecting in dev mode
-func TestInjectExtensionInstallStep_DevMode(t *testing.T) {
-	wf := &Workflow{
-		Jobs: map[string]WorkflowJob{
-			"copilot-setup-steps": {
-				Steps: []CopilotWorkflowStep{
-					{Name: "Existing step", Run: "echo test"},
-				},
-			},
-		},
-	}
-
-	err := injectExtensionInstallStep(wf, workflow.ActionModeDev, "dev")
-	if err != nil {
-		t.Fatalf("injectExtensionInstallStep() failed: %v", err)
-	}
-
-	job := wf.Jobs["copilot-setup-steps"]
-
-	// Should have 2 steps: install, existing (no verify, no checkout in dev mode)
-	if len(job.Steps) != 2 {
-		t.Fatalf("Expected 2 steps (install, existing), got %d", len(job.Steps))
-	}
-
-	// Verify install step uses curl
-	if job.Steps[0].Name != "Install gh-aw extension" {
-		t.Errorf("First step should be install, got: %s", job.Steps[0].Name)
-	}
-	if !strings.Contains(job.Steps[0].Run, "curl -fsSL") {
-		t.Errorf("Install should use curl, got: %s", job.Steps[0].Run)
-	}
-	if !strings.Contains(job.Steps[0].Run, "install-gh-aw.sh") {
-		t.Errorf("Install should reference install-gh-aw.sh, got: %s", job.Steps[0].Run)
-	}
-
-	// Verify original step is preserved
-	if job.Steps[1].Name != "Existing step" {
-		t.Errorf("Second step should be existing step, got: %s", job.Steps[1].Name)
 	}
 }
