+++ pkg/cli/git_test.go
@@ -453,190 +453,3 @@ func TestCheckWorkflowFileStatusNotInRepo(t *testing.T) {
 		t.Error("Expected empty status when not in git repository")
 	}
 }
-
-func TestToGitRootRelativePath(t *testing.T) {
-	tmpDir := testutil.TempDir(t, "test-*")
-
-	originalDir, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("Failed to get current directory: %v", err)
-	}
-	defer func() {
-		_ = os.Chdir(originalDir)
-	}()
-
-	// Create .github directory structure
-	githubDir := tmpDir + "/.github/workflows"
-	if err := os.MkdirAll(githubDir, 0755); err != nil {
-		t.Fatalf("Failed to create .github directory: %v", err)
-	}
-
-	// Create a test file in .github/workflows
-	testFile := githubDir + "/test.campaign.md"
-	if err := os.WriteFile(testFile, []byte("test"), 0644); err != nil {
-		t.Fatalf("Failed to create test file: %v", err)
-	}
-
-	// Create a subdirectory for testing from different working directories
-	subDir := tmpDir + "/pkg/cli"
-	if err := os.MkdirAll(subDir, 0755); err != nil {
-		t.Fatalf("Failed to create subdirectory: %v", err)
-	}
-
-	t.Run("from_root_directory", func(t *testing.T) {
-		if err := os.Chdir(tmpDir); err != nil {
-			t.Fatalf("Failed to change to root directory: %v", err)
-		}
-
-		result := ToGitRootRelativePath(testFile)
-		expected := ".github/workflows/test.campaign.md"
-		if result != expected {
-			t.Errorf("Expected %q, got %q", expected, result)
-		}
-	})
-
-	t.Run("from_subdirectory", func(t *testing.T) {
-		if err := os.Chdir(subDir); err != nil {
-			t.Fatalf("Failed to change to subdirectory: %v", err)
-		}
-
-		result := ToGitRootRelativePath(testFile)
-		expected := ".github/workflows/test.campaign.md"
-		if result != expected {
-			t.Errorf("Expected %q, got %q", expected, result)
-		}
-	})
-
-	t.Run("with_relative_path_from_root", func(t *testing.T) {
-		if err := os.Chdir(tmpDir); err != nil {
-			t.Fatalf("Failed to change to root directory: %v", err)
-		}
-
-		relativePath := ".github/workflows/test.campaign.md"
-		result := ToGitRootRelativePath(relativePath)
-		expected := ".github/workflows/test.campaign.md"
-		if result != expected {
-			t.Errorf("Expected %q, got %q", expected, result)
-		}
-	})
-
-	t.Run("with_relative_path_from_subdirectory", func(t *testing.T) {
-		if err := os.Chdir(subDir); err != nil {
-			t.Fatalf("Failed to change to subdirectory: %v", err)
-		}
-
-		// Relative path from subdirectory
-		relativePath := "../../.github/workflows/test.campaign.md"
-		result := ToGitRootRelativePath(relativePath)
-		expected := ".github/workflows/test.campaign.md"
-		if result != expected {
-			t.Errorf("Expected %q, got %q", expected, result)
-		}
-	})
-
-	t.Run("stability_across_directories", func(t *testing.T) {
-		// Test from root
-		if err := os.Chdir(tmpDir); err != nil {
-			t.Fatalf("Failed to change to root directory: %v", err)
-		}
-		result1 := ToGitRootRelativePath(testFile)
-
-		// Test from subdirectory
-		if err := os.Chdir(subDir); err != nil {
-			t.Fatalf("Failed to change to subdirectory: %v", err)
-		}
-		result2 := ToGitRootRelativePath(testFile)
-
-		// Results should be identical
-		if result1 != result2 {
-			t.Errorf("Path not stable across directories: from root=%q, from subdir=%q", result1, result2)
-		}
-
-		expected := ".github/workflows/test.campaign.md"
-		if result1 != expected {
-			t.Errorf("Expected %q, got %q", expected, result1)
-		}
-	})
-}
-
-func TestToGitRootRelativePathWithoutGithubDir(t *testing.T) {
-	tmpDir := testutil.TempDir(t, "test-*")
-
-	originalDir, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("Failed to get current directory: %v", err)
-	}
-	defer func() {
-		_ = os.Chdir(originalDir)
-	}()
-
-	// Create a subdirectory but NO .github directory
-	subDir := tmpDir + "/subdir"
-	if err := os.MkdirAll(subDir, 0755); err != nil {
-		t.Fatalf("Failed to create subdirectory: %v", err)
-	}
-
-	testFile := subDir + "/test.md"
-	if err := os.WriteFile(testFile, []byte("test"), 0644); err != nil {
-		t.Fatalf("Failed to create test file: %v", err)
-	}
-
-	if err := os.Chdir(tmpDir); err != nil {
-		t.Fatalf("Failed to change to temp directory: %v", err)
-	}
-
-	// Should fall back to relative path from current working directory
-	result := ToGitRootRelativePath(testFile)
-
-	// Should return a relative path (not absolute)
-	if strings.HasPrefix(result, "/") || (len(result) > 2 && result[1] == ':') {
-		t.Errorf("Expected relative path fallback, got absolute path: %q", result)
-	}
-
-	// Should contain the filename
-	if !strings.Contains(result, "test.md") {
-		t.Errorf("Expected result to contain filename, got: %q", result)
-	}
-}
-
-func TestToGitRootRelativePathWithNestedGithubDir(t *testing.T) {
-	tmpDir := testutil.TempDir(t, "test-*")
-
-	originalDir, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("Failed to get current directory: %v", err)
-	}
-	defer func() {
-		_ = os.Chdir(originalDir)
-	}()
-
-	// Create nested structure: root/.github and root/subproject/.github
-	rootGithubDir := tmpDir + "/.github/workflows"
-	if err := os.MkdirAll(rootGithubDir, 0755); err != nil {
-		t.Fatalf("Failed to create root .github directory: %v", err)
-	}
-
-	subprojectGithubDir := tmpDir + "/subproject/.github/workflows"
-	if err := os.MkdirAll(subprojectGithubDir, 0755); err != nil {
-		t.Fatalf("Failed to create subproject .github directory: %v", err)
-	}
-
-	// Create test file in subproject
-	testFile := subprojectGithubDir + "/test.md"
-	if err := os.WriteFile(testFile, []byte("test"), 0644); err != nil {
-		t.Fatalf("Failed to create test file: %v", err)
-	}
-
-	// Change to subproject directory
-	subprojectDir := tmpDir + "/subproject"
-	if err := os.Chdir(subprojectDir); err != nil {
-		t.Fatalf("Failed to change to subproject directory: %v", err)
-	}
-
-	// Should find the closest .github directory (in subproject)
-	result := ToGitRootRelativePath(testFile)
-	expected := ".github/workflows/test.md"
-	if result != expected {
-		t.Errorf("Expected %q, got %q (should use closest .github)", expected, result)
-	}
-}
