+++ pkg/cli/mcp_registry_live_test.go
@@ -107,85 +107,6 @@ func TestMCPRegistryClient_LiveSearchServers(t *testing.T) {
 	})
 }
 
-// TestMCPRegistryClient_LiveGetServer tests GetServer against the live GitHub MCP registry
-func TestMCPRegistryClient_LiveGetServer(t *testing.T) {
-	// Create client with default production registry URL
-	client := NewMCPRegistryClient(string(constants.DefaultMCPRegistryURL))
-
-	// Test 1: Get a known server that should exist in the registry
-	// We'll use the GitHub MCP server as it's maintained by GitHub
-	t.Run("get_github_server", func(t *testing.T) {
-		// First, search for GitHub servers to find an actual server name
-		servers, err := client.SearchServers("github")
-		if err != nil {
-			if strings.Contains(err.Error(), "network") || strings.Contains(err.Error(), "firewall") ||
-				strings.Contains(err.Error(), "403") || strings.Contains(err.Error(), "connection") {
-				t.Skipf("Skipping due to network restrictions: %v", err)
-				return
-			}
-			t.Fatalf("SearchServers failed: %v", err)
-		}
-
-		if len(servers) == 0 {
-			t.Skip("No GitHub servers found in registry to test GetServer")
-			return
-		}
-
-		// Get the first server's name
-		serverName := servers[0].Name
-		t.Logf("Testing GetServer with: %s", serverName)
-
-		// Now test GetServer with that name
-		server, err := client.GetServer(serverName)
-		if err != nil {
-			if strings.Contains(err.Error(), "network") || strings.Contains(err.Error(), "firewall") ||
-				strings.Contains(err.Error(), "403") || strings.Contains(err.Error(), "connection") ||
-				strings.Contains(err.Error(), "503") || strings.Contains(err.Error(), "502") ||
-				strings.Contains(err.Error(), "upstream") || strings.Contains(err.Error(), "reset") {
-				t.Skipf("Skipping due to registry unavailability: %v", err)
-				return
-			}
-			t.Fatalf("GetServer failed for '%s': %v", serverName, err)
-		}
-
-		// Validate the returned server
-		if server.Name != serverName {
-			t.Errorf("Expected server name '%s', got '%s'", serverName, server.Name)
-		}
-		if server.Description == "" {
-			t.Logf("Warning: Server '%s' has empty description (field may be optional)", serverName)
-		}
-		if server.Transport == "" {
-			t.Errorf("Server transport is empty")
-		}
-
-		t.Logf("✓ Successfully retrieved server: %s", server.Name)
-		t.Logf("  Description: %s", server.Description)
-		t.Logf("  Transport: %s", server.Transport)
-		if server.Command != "" {
-			t.Logf("  Command: %s", server.Command)
-		}
-		if server.RuntimeHint != "" {
-			t.Logf("  Runtime Hint: %s", server.RuntimeHint)
-		}
-	})
-
-	// Test 2: Get a server that doesn't exist
-	t.Run("get_nonexistent_server", func(t *testing.T) {
-		_, err := client.GetServer("nonexistent/fake-server-12345")
-		if err == nil {
-			t.Errorf("Expected error for nonexistent server, got nil")
-		}
-
-		expectedErrorSubstring := "not found in registry"
-		if !strings.Contains(err.Error(), expectedErrorSubstring) {
-			t.Errorf("Expected error to contain '%s', got: %s", expectedErrorSubstring, err.Error())
-		}
-
-		t.Logf("✓ Correctly returned error for nonexistent server")
-	})
-}
-
 // TestMCPRegistryClient_LiveResponseStructure tests that the v0.1 API structure is correctly parsed
 func TestMCPRegistryClient_LiveResponseStructure(t *testing.T) {
 	// Create client with default production registry URL
