+++ pkg/cli/mcp_registry_test.go
@@ -101,112 +101,6 @@ func TestMCPRegistryClient_SearchServers(t *testing.T) {
 	}
 }
 
-func TestMCPRegistryClient_GetServer(t *testing.T) {
-	// Create a test server that mocks the MCP registry API
-	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		expectedPath := "/servers"
-		if r.URL.Path != expectedPath {
-			t.Errorf("Expected path %s, got %s", expectedPath, r.URL.Path)
-		}
-
-		// No longer check for search query parameter since we now fetch all servers and filter locally
-
-		// Return mock response with v0.1 structure based on official specification
-		response := `{
-			"servers": [
-				{
-					"server": {
-						"name": "io.github.makenotion/notion-mcp-server",
-						"description": "MCP server for Notion integration",
-						"version": "1.0.0",
-						"repository": {
-							"url": "https://github.com/example/notion-mcp",
-							"source": "github"
-						},
-						"packages": [
-							{
-								"registryType": "npm",
-								"identifier": "notion-mcp",
-								"version": "1.0.0",
-								"runtimeHint": "node",
-								"transport": {
-									"type": "stdio"
-								},
-								"packageArguments": [
-									{
-										"type": "positional",
-										"value": "notion-mcp"
-									}
-								],
-								"environmentVariables": [
-									{
-										"name": "NOTION_TOKEN",
-										"description": "Notion API token",
-										"isRequired": true,
-										"isSecret": true
-									}
-								]
-							}
-						]
-					},
-					"_meta": {
-						"io.modelcontextprotocol.registry/official": {
-							"status": "active",
-							"publishedAt": "2025-01-01T10:30:00Z",
-							"isLatest": true
-						}
-					}
-				}
-			]
-		}`
-
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(response))
-	}))
-	defer testServer.Close()
-
-	// Create client with test server URL
-	client := NewMCPRegistryClient(testServer.URL)
-
-	// Test get server
-	serverInfo, err := client.GetServer("io.github.makenotion/notion-mcp-server")
-	if err != nil {
-		t.Fatalf("GetServer failed: %v", err)
-	}
-
-	if serverInfo.Name != "io.github.makenotion/notion-mcp-server" {
-		t.Errorf("Expected server name 'io.github.makenotion/notion-mcp-server', got '%s'", serverInfo.Name)
-	}
-}
-
-func TestMCPRegistryClient_GetServerNotFound(t *testing.T) {
-	// Create a test server that returns empty response (no matching servers)
-	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		response := `{
-			"servers": []
-		}`
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(response))
-	}))
-	defer testServer.Close()
-
-	// Create client with test server URL
-	client := NewMCPRegistryClient(testServer.URL)
-
-	// Test get server that doesn't exist
-	_, err := client.GetServer("nonexistent")
-	if err == nil {
-		t.Fatal("Expected error for nonexistent server, got nil")
-	}
-
-	expectedError := "MCP server 'nonexistent' not found in registry"
-	if err.Error() != expectedError {
-		t.Errorf("Expected error '%s', got '%s'", expectedError, err.Error())
-	}
-}
-
 func TestNewMCPRegistryClient_DefaultURL(t *testing.T) {
 	client := NewMCPRegistryClient("")
 	expectedURL := "https://api.mcp.github.com/v0.1"
