+++ pkg/cli/spec_test.go
@@ -571,112 +571,6 @@ func TestParseSourceSpec(t *testing.T) {
 	}
 }
 
-func TestBuildSourceString(t *testing.T) {
-	tests := []struct {
-		name     string
-		workflow *WorkflowSpec
-		expected string
-	}{
-		{
-			name: "workflow with version",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "owner/repo",
-					Version:  "v1.0.0",
-				},
-				WorkflowPath: "workflows/ci-doctor.md",
-			},
-			expected: "owner/repo/workflows/ci-doctor.md@v1.0.0",
-		},
-		{
-			name: "workflow with branch",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "owner/repo",
-					Version:  "main",
-				},
-				WorkflowPath: "workflows/helper.md",
-			},
-			expected: "owner/repo/workflows/helper.md@main",
-		},
-		{
-			name: "workflow without version",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "owner/repo",
-					Version:  "",
-				},
-				WorkflowPath: "workflows/test.md",
-			},
-			expected: "owner/repo/workflows/test.md",
-		},
-		{
-			name: "workflow with nested path",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "owner/repo",
-					Version:  "v2.0.0",
-				},
-				WorkflowPath: "path/to/workflow.md",
-			},
-			expected: "owner/repo/path/to/workflow.md@v2.0.0",
-		},
-		{
-			name: "empty repo",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "",
-					Version:  "v1.0.0",
-				},
-				WorkflowPath: "workflows/test.md",
-			},
-			expected: "",
-		},
-		{
-			name: "empty workflow path",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "owner/repo",
-					Version:  "v1.0.0",
-				},
-				WorkflowPath: "",
-			},
-			expected: "",
-		},
-		{
-			name: "local workflow without version",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "owner/repo",
-					Version:  "",
-				},
-				WorkflowPath: "./workflows/local.md",
-			},
-			expected: "owner/repo/workflows/local.md",
-		},
-		{
-			name: "local workflow with version (should still work)",
-			workflow: &WorkflowSpec{
-				RepoSpec: RepoSpec{
-					RepoSlug: "owner/repo",
-					Version:  "v1.0.0",
-				},
-				WorkflowPath: "./test.md",
-			},
-			expected: "owner/repo/test.md@v1.0.0",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := buildSourceString(tt.workflow)
-			if result != tt.expected {
-				t.Errorf("buildSourceString() = %v, want %v", result, tt.expected)
-			}
-		})
-	}
-}
-
 func TestBuildSourceStringWithCommitSHA(t *testing.T) {
 	tests := []struct {
 		name      string
